Add files via upload
The final course was a Capstone project that encompassed most of the classes we took. We were given a dataset about restaurants that contained names, cuisine types, address, neighborhood, average prices, ratings, the type of restaurant, comments from social media, and many other variables, all from the Bangalore, India area. This area is fast growing and needs more dining options. After cleaning the data and doing some descriptive analytics, we were asked to scope out a project plan for this data. I chose to look for areas that had a demand for a certain cuisine of food, but yet the quality of that food was low. With that information, I found small restaurants that had a high quality that could perhaps be talked into franchising by my client, Betty Bigbucks. All analysis done with R Markdown and Excel Solver, using various regression techniques, predictive analytics, and linear programming.